@@ -1,19 +1,15 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { AddComponent } from './add/add.component';
-import { ContactComponent } from './contact/contact.component';
-import { EditComponent } from './edit/edit.component';
-import { HomeComponent } from './home/home.component';
+import { AboutComponent } from './about/about.component';
 import { ViewComponent } from './view/view.component';
 
 
 const routes: Routes = [
-  { path: 'home', component: HomeComponent },
-  { path: '', component: HomeComponent },
+  { path: '', component: AboutComponent },
   { path: 'view', component: ViewComponent },
-  { path: 'edit', component: EditComponent },
   { path: 'add', component: AddComponent },
-  { path: 'contactUs', component: ContactComponent }
+  { path: 'home', component: AboutComponent }
 ];
 
 @NgModule({
