@@ -2,5 +2,5 @@ export class Constants{
 
   public static addArtWorkObjURL:string='http://ip.jsontest.com';
   public static searchArtWorkURL:string='http://ip.jsontest.com';
-
+  public static latestArtworksURL:string='http://ip.jsontest.com';
 }
