@@ -1,23 +1,47 @@
 import { Component, OnInit } from '@angular/core';
+import { CrudService } from 'src/app/services/crud.service';
 
 @Component({
   selector: 'app-view',
-  templateUrl: './view.component.html',
-  styleUrls: ['./view.component.css']
+  templateUrl: './view.component.html'
 })
 export class ViewComponent implements OnInit {
 
-  constructor() { }
+  constructor(private service: CrudService) { }
 
   ngOnInit(): void {
+      this.search();//ewali remove this
   }
-  pictureData:Array<any> = [
-    {imageLoc:'assets/images/van-gogh.jpg', title:'Starry Night', desc:'This is Starry Night by Van Gogh'},
-    {imageLoc:'assets/images/edwardhopper.jpg', title:'Light House', desc:'Light House by Ed hopper'},
-    {imageLoc:'assets/images/van-gogh2.jpg', title:'Landscape', desc:'Landscape by Van Gogh'},
-    {imageLoc:'assets/images/Sunrise.jpg', title:'Sunrise', desc:'Sunrise by Sunrise'},
-    {imageLoc:'assets/images/portfolio-05.jpg', title:'Fifth Awesome', desc:'Ut sollicitudin risus'},
-    {imageLoc:'https://upload.wikimedia.org/wikipedia/commons/thumb/e/ea/Van_Gogh_-_Starry_Night_-_Google_Art_Project.jpg/757px-Van_Gogh_-_Starry_Night_-_Google_Art_Project.jpg', title:'Landscape', desc:'Caspar_David_Friedrich'}];
+  artworks:Array<any> = [];
+
+  type_values:any[] = [{label:'Artist Name', value:'artistName'},
+       {label:'Art Name', value:'artName'},
+       {label:'Art category', value:'artCat'},
+       {label:'Creation Year', value:'creationYear'},
+       {label:'Acquisition Year', value:'acquistionYear'},
+       {label:'Color', value:'color'},
+       {label:'Culture', value:'culture'},
+       {label:'Category', value:'category'},
+       {label:'Medium', value:'medium'},
+       {label:'BRG', value:'brg'}
+    ];
+
+  searchRequest:any = {type: this.type_values[0].value, value:''};
+
+  search():void{
+    /* if(this.searchRequest.value==''){
+      alert('Please enter some search value');
+      return;
+    } */
+    this.service.searchArtWork(this.searchRequest).subscribe(response => {
+            let res= response;
+        this.service.getLatestArtWorks().subscribe(response => {
+                   this.artworks = response;
+        });
+  });
+
+  }
+
 
 
 
